move DownloadGroth16ArtifactTask to dedicated kotlin file
