+++ cryptography/hedera-cryptography-rpm/build.gradle.kts
@@ -5,11 +5,11 @@ import org.gradle.kotlin.dsl.withType
 import org.hiero.gradle.environment.EnvAccess
 import org.hiero.gradle.tasks.CargoBuildTask
 
-// SPDX-License-Identifier: Apache-2.0
 plugins {
     id("org.hiero.gradle.module.library")
     id("org.hiero.gradle.feature.rust")
     id("org.hiero.gradle.feature.test-multios")
+    id("DownloadGroth16ArtifactTask")
 }
 
 cargo { libname = "raps" }
@@ -366,53 +366,6 @@ testModuleInfo {
     requires("org.junit.jupiter.params")
 }
 
-// Support the ProofCompressorTest by preparing a large binary artifact that the compressor uses:
-val groth16ArtifactDir = layout.buildDirectory.dir("groth16-artifact")
-
-abstract class DownloadGroth16ArtifactTask : DefaultTask() {
-    @get:Inject protected abstract val files: FileOperations
-
-    @get:OutputDirectories abstract val groth16Dir: DirectoryProperty
-
-    @TaskAction
-    fun action() {
-        val out = groth16Dir.get().dir("v5.0.0")
-        files.mkdir(out)
-
-        // This is a 3GB download, so we only do this if we must:
-        val filename = "v5.0.0-groth16.tar.gz"
-        val uri = "https://builds.hedera.com/tss/sp1/groth16/v5.0/$filename"
-        val url = URI(uri).toURL()
-        val tarball = groth16Dir.get().file(filename).asFile
-        if (!tarball.exists()) {
-            println("Downloading $uri to ${tarball.absolutePath}")
-            // file.writeBytes(url.readBytes()) runs out of heap space, so we copy streams instead:
-            url.openStream().use { input ->
-                tarball.outputStream().use { output -> input.copyTo(output) }
-            }
-        } else {
-            println("$uri has already been downloaded as: ${tarball.absolutePath}")
-        }
-        // Just one of the artifact files, good enough for a quick test:
-        val testArtifactFileName = "groth16_circuit.bin"
-        if (!files.file(out.file(testArtifactFileName)).exists()) {
-            println("Extracting ${tarball.absolutePath} to ${out.asFile.absolutePath}")
-            files.sync {
-                from(files.tarTree(tarball))
-                into(out)
-            }
-        } else {
-            println(
-                "Not extracting Groth16 artifact as it already exists: e.g. ${out.file(testArtifactFileName).asFile.absolutePath}"
-            )
-        }
-    }
-}
-
-tasks.register<DownloadGroth16ArtifactTask>("downloadGroth16ArtifactTask") {
-    groth16Dir.convention(groth16ArtifactDir)
-}
-
 tasks.test {
     dependsOn("downloadGroth16ArtifactTask")
     environment(
@@ -421,8 +374,13 @@ tasks.test {
             "TSS_LIB_NUM_OF_CORES" to "10",
             // For the compressor that runs Go:
             "GOMAXPROCS" to "10",
-            // Path to Groth16 artifacts. This depends on the downloadGroth16ArtifactTask :
-            "SP1_GROTH16_CIRCUIT_PATH" to groth16ArtifactDir.get().asFile.absolutePath,
+            // Path to Groth16 artifacts:
+            "SP1_GROTH16_CIRCUIT_PATH" to
+                (tasks.named("downloadGroth16ArtifactTask").get().property("groth16Dir")
+                        as DirectoryProperty)
+                    .get()
+                    .asFile
+                    .absolutePath,
         )
     )
 }
